requirements.txt
playwright
requests
python-telegram-bot==13.7